requirements.txt
# 爬虫框架
scrapy>=2.6.0
# 数据库
pymysql>=1.0.2
SQLAlchemy>=1.4.40
# 网络请求
requests>=2.28.1
# HTML解析
beautifulsoup4>=4.11.1
lxml>=4.9.1
# 图片处理
Pillow>=9.2.0
# 浏览器自动化
selenium>=4.4.0
webdriver-manager>=3.8.3
# 日期处理
python-dateutil>=2.8.2
# 并发处理
concurrent-futures>=3.0.5
# 日志
loguru>=0.6.0
# 工具
tqdm>=4.64.0
# 测试
pytest>=7.1.3
# 数据处理
pandas>=1.4.3
numpy>=1.23.2
# 定时任务
schedule>=1.1.0
# 配置管理
python-dotenv>=0.21.0 